made shit and stuf

+++ src/Point.java
@@ -1,17 +1,17 @@
 public class Point {
-    private float x;
-    private float y;
+    private double x;
+    private double y;
 
-    public Point(float x, float y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
     public void printer() {
